pyproject.toml
[tool.poetry]
name = "product-shop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
drf-yasg = "^1.21.8"
django-filter = "^24.3"
setuptools = "^75.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
